moved added mobile folder to nav
